Revert to older version of pkg/browser
pkg/browser#26 added support for opening URLs in text-based browsers,
such as lynx. Unfortunately this does not play well with the CLI's
`linkerd-buoyant install | kubectl apply -f -` pattern, as we intend for
only a k8s manifest to be piped to kubectl.

Revert to an older version of `pkg/browser` that only supports
`xdg-open`.

Fixes #3

Signed-off-by: Andrew Seigner <siggy@buoyant.io>