Extract CP roots from config map (#25)
When running with the latest Linkerd edge bloud agent is sending a certificate message with the Roots being empty. The reason for that is that in the bcloud agent, we look for the `EnvVar.Value` field rather than the `EnvVar.ValueFrom`. In https://github.com/linkerd/linkerd2/pull/6455 the proxy manifest was changed to load the trust anchors from a config map. This causes the described above problem. 

This can be fixed by making the agent consider this value when inspecting the env vars. It is important to make this backwards compatible - check the `Value` field and only if not present, check the `ValueFrom` one. This might require RBAC changes.

Note to reviewers: 

In order to test, run this agent with the latest stable to ensure the change is backwards compatible. After that you can run it with the latest edge to convince yourself that the problem has been fixed. In both cases you need to make sure that the certificates in the control plane page are populated.

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>