Move the CLI to the new auth flow (#51)
This PR changes the CLI to work by using the new auth/autoregistration flow. This means that when agent registration needs to happen, the CLI will not bring you to the Bcloud UI to create the agent but will instead render a manifest that will take care of registration. 

Additionally, the `check` command has been tweaked to check for the presence of the resources that are not required.